requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
Werkzeug==3.0.1
requests==2.31.0
google-generativeai==0.3.2
PyPDF2==3.0.1
numpy==1.26.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
openai==1.3.7
python-dotenv==1.0.0
authlib==1.3.0
pytz==2024.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
eventlet==0.33.3